Changed the way combat works & fixed respawn crashes
New Combat: Red = Attack, Blue = Defense, Green = Speed
Respawn fix: There was an issue that when players were killed while in a spawnpoint, the spawnpoint would never recieve a TriggerLeave event.  I have fixed this by adding players to a spawnpoint's playerList on enter and removing them on leave and when a player is killed.